Move everything back into the main executable.
Why? Because Cargo optimizes too quickly. When the library and the executable are in different
places, then cargo will compile one after the other. Makes sense, right? Yes, but it will now
compile the small C file into the library, and when optimizing the code it will throw out
important parts of the file, because they are only used after it is linked with the main
executable.

Another workaround would have been to move everything to Rust, but I didn't manage (yet!)
to get all the code right.